add plugin to delete old gatling report

+++ clean_old_gatling_reports.sh
@@ -3,6 +3,11 @@
 # Define the Gatling results directory
 RESULTS_DIR="target/gatling"
 
-# Find all directories in the Gatling results directory, sort them by modification time, and skip the most recent one
-cd $RESULTS_DIR
-ls -dt */ | tail -n +2 | xargs rm -rf
+# Check if the directory exists
+if [ -d "$RESULTS_DIR" ]; then
+  cd $RESULTS_DIR
+  # Find all directories, sort them by modification time, and skip the most recent one
+  ls -dt */ | tail -n +2 | xargs rm -rf
+else
+  echo "Directory $RESULTS_DIR does not exist."
+fi
